Comment out checks that are causing failures on ^/stable/11, post-r337133
Some usr.bin/procstat code has not been MFCed to ^/stable/11 for command line
handling, and unfortunately, r337133 (which was MFCed as r337542) relies on
that support.

Comment out the checks and add a pointer to the PR for code archaeology and
to cause future potential parties to verify changes when backporting them.

This is a direct commit to ^/stable/11

PR:		237445
Reviewed by:	markj
Approved by:	emaste (mentor; implicit)
Differential Revision: https://reviews.freebsd.org/D19998


git-svn-id: https://svn.freebsd.org/base/stable/11@346515 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f