Fix mouse cursor coloring in depths > 8 (previously, a hack that only worked right for white interiors and black borders was used).  Advertise this by changing the default colors to a red interior and a white border (the same as the kernel default).  Add undocumented env variables for changing these colors.  Also change to the larger and better-shaped 16x10 cursor sometimes used in the kernel.  The kernel choice is fancier, but libvgl is closer to supporting the larger cursors needed in newer modes.
The (n)and-or logic for the cursor doesn't work right for more than 2
colors.  The (n)and part only masks out all color bits for the pixel
under the cursor when all bits are set in the And mask.  With more
complicated logic, the non-masked bits could be used to implement
translucent cursors, but they actually just gave strange colors
(especially in packed and planar modes where the bits are indirect
through 1 or 2 palettes so it is hard to predict the final color).
They also gave a bug for writing pixels under the cursor.  The
non-masked bits under the cursor were not combined in this case.

Drop support for combining with bits under the cursor by making any nonzero
value in the And mask mean all bits set.

Convert the Or mask (which is represented as a half-initialized 256-color
bitmap) to a fully initialized bitmap with the correct number of colors.
The 256-color representation must be as in 3:3:2 direct mode iff the final
bitmap has more than 256 colors.  The conversion of colors is not very
efficient, so convert at initialization time.


git-svn-id: https://svn.freebsd.org/base/head@346580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f