MFC r333322: Keep CARP state as INIT when net.inet.carp.allow=0.
Currently when net.inet.carp.allow=0 CARP state remains as MASTER, which is
not very useful (if there are other masters -- it can lead to split brain,
if there are none -- it makes no sense).  Having it as INIT makes it clear
that carp packets are disabled.

Sponsored by:	iXsystems, Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@346770 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f