Add a note to MAINTAINERS for lkpi for graphics
Add a note to MAINTAINERS requesting pre-commit review from the graphics
team, using phabricator, for changes to the lkpi subsystem.  This is done in
order to give us a chance to test the graphics drivers (drm drivers) for
regressions, and to try to avoid breakage, errors and issues with the
graphics drivers.
The review is done via the #x11 group on phabricator.

Please note that hselasky also want to review changes.

Discussed with:	hselasky, imp
Approved by:	imp


git-svn-id: https://svn.freebsd.org/base/head@346899 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f