MFC r346539:
Fix `get_int_via_sysctlbyname(..)` on Jenkins

Initialize `oldlen` to the size of the value, instead of leaving the value
unitialized. Leaving it unitialized seems to work by accident on amd64 when
running 64-bit programs, but not on i386.

This matches patterns in use in other programs.

PR:		237458
Tested on:	^/head (amd64), ^/stable/11 (i386)


git-svn-id: https://svn.freebsd.org/base/stable/11@346911 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f