MFC r346489:
Move the reporting of spurious interrupts under bootverbose control, because
occasional spurious interrupts are a normal thing on this hardware.  Also,
change the name of the cpu-local interrupt controller driver from local_intc
to lintc, because the name gets built into interrupt names, which have to
fit into a 19-byte field for stats reporting (so this allows 5 more bytes
of the actual interrupt name to be displayed).


git-svn-id: https://svn.freebsd.org/base/stable/12@346937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f