MFC r346567:
cxgbe(4): Make sure bundled_fw is always initialized before use.

This fixes a bug that prevented the driver from auto-flashing the
firmware when it didn't see one on the card.  This feature was
introduced in r321390 and this bug was introduced in r343269.

Reported by:	gallatin@
Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/12@346975 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f