MFC 325727: bhyve: avoid applying capsicum capabilities to file that was not opened
When using -l option targeting file that can't be opened (ie. nmdm module
is not loaded and /dev/nmdm* is specified) bhyve tries to apply capsicum
capabilities to a file that was not opened.

Enclose that code in an if statement and only run it on correctly opened
descriptor also providing meaningful message in case of an error.


git-svn-id: https://svn.freebsd.org/base/stable/11@347036 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f