MFC r336937: Send consistent SEG.WIN when using timewait codepath for TCP.
When sending TCP segments from the timewait code path, a stored
value of the last sent window is used. Use the same code for
computing this in the timewait code path as in the main code
path used in tcp_output() to avoid inconsistencies.

MFC r344148:
Fix a byte ordering issue for the advertised receiver window in ACK
segments sent in TIMEWAIT state, which I introduced in r336937.


git-svn-id: https://svn.freebsd.org/base/stable/11@347161 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f