MFC r345303, 345658, and partially MFC r345305
These are:
r345303: prevent possible infinite loop in iflib_encap
r345305: expose the Rx mbuf buffer size to drivers
r345658: return ENETDOWN when the network device is down

r345305 is only partially MFC'd with no mergeinfo because it includes
changes to iflib-using drivers, and none of the drivers it changes use
iflib in stable/11. This commit just makes the function it adds available
for use with iflib-using kernel modules.

Sponsored by:	Intel Corporation


git-svn-id: https://svn.freebsd.org/base/stable/11@347212 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f