MFC r347365, r347703
r347365:
Update dme(4) to reflect that it will not be removed due to FCP-101.

dme(4) is the built-in NIC on a couple non-expandable mips platforms and
thus should remain.  The FCP has been updated to reflect this fact.

Discussed with:	imp

r347703:
FCP-101: ae(4) is sufficently popular to be moved to the keep list.


git-svn-id: https://svn.freebsd.org/base/stable/12@347848 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f