MFC r346652-r346654
r346652:
Complain and exit the script if the 'make install' phase fails.  Also,
there is no need to install any debug files.

r346653:
The zfs module has grown a dependency on zcl_nfs4.ko, so copy it into all
the test images.

r346654:
For the geli-gpt-zfs test images, both bios and uefi flavors, add a dummy
ufs partition as p2, and put the zfs partition at p3, to test the ability
of the zfs probe code to find a zfs pool on something other than the first
partition.


git-svn-id: https://svn.freebsd.org/base/stable/12@348145 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f