Add warning that the PNP info has to follow the module declaration.
Due to how the linker.hints file is laid out, we'll associate the pnp
info with the wrong module if the module declaration comes after the
pnp info. Until that limiation is removed, we need to have this
ordering. Ideally, we'd also enforce the ordering somehow, but I've
come up with no way to do that yet...


git-svn-id: https://svn.freebsd.org/base/head@348171 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f